Split the extension list function into a separate file.
